pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-kafka-streaming-pipeline"
version = "0.1.0"
description = "Real-time crypto trade streaming pipeline: Binance → Kafka → Parquet → Teradata → Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.3",
    "websockets>=12.0",
    "orjson>=3.9",
    "jsonschema>=4.21",
    "polars>=0.20",
    "teradatasql>=17.20",
    "structlog>=24.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "dbt-core>=1.7",
    "dbt-teradata>=1.7",
    "boto3>=1.34",
    "pandas>=2.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.30",
    "plotly>=5.18",
]

[project.scripts]
crypto-producer  = "crypto_pipeline.producer.main:main"
crypto-consumer  = "crypto_pipeline.consumer.main:main"
crypto-init-db   = "crypto_pipeline._cli:init_db"
crypto-query-db  = "crypto_pipeline._cli:query_db"
crypto-dbt       = "crypto_pipeline._cli:run_dbt"
crypto-dashboard = "crypto_pipeline._cli:dashboard"

[tool.hatch.build.targets.wheel]
packages = ["src/crypto_pipeline"]

# Bundle dashboard/app.py into the installed package so `uvx crypto-dashboard`
# can locate it regardless of the working directory's layout.
[tool.hatch.build.targets.wheel.force-include]
"dashboard/app.py" = "crypto_pipeline/dashboard_app.py"
